V1.2 (Se agrego productos)

+++ src/main/java/com/unla/grupo24/repositories/RolUsuarioRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.unla.grupo24.entities.RolUsuario;
 
-public interface RolUsuarioRepository extends JpaRepository<RolUsuario, Long> {
+public interface RolUsuarioRepository extends JpaRepository<RolUsuario, Integer> {
 }
